Extract the AppBar into NavBar
